stac_browser/requirements.txt
aws-cdk-lib<3.0.0,>=2.15.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws_apigatewayv2_alpha~=2.47.0.a0
aws_cdk.aws_apigatewayv2_integrations_alpha~=2.47.0.a0
eoapi-cdk==5.4.0